requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
isort==5.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
line-bot-sdk==1.17.0
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.6.0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
